% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTensor_Misc.R
\docType{data}
\name{faces_tnsr}
\alias{faces_tnsr}
\title{ORL Database of Faces}
\format{A Tensor object with modes 92 x 112 x 40 x 10. The first two modes correspond to the image pixels, the third mode corresponds to the individual, and the last mode correpsonds to the lighting.}
\source{
\url{http://www.cl.cam.ac.uk/research/dtg/attarchive/facedatabase.html}
}
\usage{
faces_tnsr
}
\description{
A dataset containing pictures of 40 individuals under 10 different lightings. Each image has 92 x 112 pixels. Structured as a 4-tensor with modes 92 x 112 x 40 x 10.
}
\seealso{
\code{\link{plot_orl}}
}
\keyword{datasets}
