% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTensor_Misc.R
\name{hadamard_list}
\alias{hadamard_list}
\title{List hadamard Product}
\usage{
hadamard_list(L)
}
\arguments{
\item{L}{list of matrices or vectors}
}
\value{
matrix that is the hadamard product
}
\description{
Returns the hadamard (element-wise) product from a list of matrices or vectors. Commonly used for n-mode products and various Tensor decompositions.
}
\note{
The modes/dimensions of each element in the list must match.
}
\examples{
lizt <- list('mat1' = matrix(runif(40),ncol=4), 
'mat2' = matrix(runif(40),ncol=4),
'mat3' = matrix(runif(40),ncol=4))
dim(hadamard_list(lizt))
}
\seealso{
\code{\link{kronecker_list}}, \code{\link{khatri_rao_list}}
}
