% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hamadard_list}
\alias{hamadard_list}
\title{List Hamadard Product}
\usage{
hamadard_list(L)
}
\arguments{
\item{L}{list of matrices or vectors}
}
\value{
matrix that is the Hamadard product
}
\description{
Returns the Hamadard (element-wise) product from a list of matrices or vectors. Commonly used for n-mode products and various Tensor decompositions.
}
\note{
The modes/dimensions of each element in the list must match.
}
\examples{
lizt <- list('mat1' = matrix(runif(40),ncol=4),
'mat2' = matrix(runif(40),ncol=4),
'mat3' = matrix(runif(40),ncol=4))
dim(hamadard_list(lizt))
}
\seealso{
\code{\link{kronecker_list}}, \code{\link{khatri_rao_list}}
}

