% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{initialize-methods}
\alias{initialize,Tensor-method}
\alias{initialize-methods}
\title{Initializes a Tensor instance}
\usage{
\S4method{initialize}{Tensor}(.Object, num_modes = NULL, modes = NULL,
  data = NULL)
}
\arguments{
\item{.Object}{the tensor object}

\item{num_modes}{number of modes of the tensor}

\item{modes}{modes of the tensor}

\item{data}{can be vector, matrix, or array}
}
\description{
Not designed to be called by the user. Use \code{as.tensor} instead.
}
\seealso{
\code{as.tensor}
}

