% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTensor_Class.R
\docType{methods}
\name{innerProd-methods}
\alias{innerProd}
\alias{innerProd,Tensor,Tensor-method}
\alias{innerProd-methods}
\title{Tensors Inner Product}
\usage{
innerProd(tnsr1, tnsr2)

\S4method{innerProd}{Tensor,Tensor}(tnsr1, tnsr2)
}
\arguments{
\item{tnsr1}{first Tensor instance}

\item{tnsr2}{second Tensor instance}
}
\value{
inner product between \code{x1} and \code{x2}
}
\description{
Returns the inner product between two Tensors
}
\details{
\code{innerProd(tnsr1,tnsr2)}
}
\examples{
tnsr1 <- rand_tensor()
tnsr2 <- rand_tensor()
innerProd(tnsr1,tnsr2)
}

