% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTensor_Misc.R
\name{khatri_rao}
\alias{khatri_rao}
\title{Khatri-Rao Product}
\usage{
khatri_rao(x, y)
}
\arguments{
\item{x}{first matrix}

\item{y}{second matrix}
}
\value{
matrix that is the Khatri-Rao product
}
\description{
Returns the Khatri-Rao (column-wise Kronecker) product of two matrices. If the inputs are vectors then this is the same as the Kronecker product.
}
\note{
The number of columns must match in the two inputs.
}
\examples{
dim(khatri_rao(matrix(runif(12),ncol=4),matrix(runif(12),ncol=4)))
}
\seealso{
\code{\link{kronecker}}, \code{\link{khatri_rao_list}}
}

