% Generated by roxygen2 (4.0.2): do not edit by hand
\name{kronecker_list}
\alias{kronecker_list}
\title{List Kronecker Product}
\usage{
kronecker_list(L)
}
\arguments{
\item{L}{list of matrices or vectors}
}
\value{
matrix that is the Kronecker product
}
\description{
Returns the Kronecker product from a list of matrices or vectors. Commonly used for n-mode products and various Tensor decompositions.
}
\examples{
smalllizt <- list('mat1' = matrix(runif(12),ncol=4),
'mat2' = matrix(runif(12),ncol=4),
'mat3' = matrix(runif(12),ncol=4))
dim(kronecker_list(smalllizt))
}
\seealso{
\code{\link{hamadard_list}}, \code{\link{khatri_rao_list}}, \code{\link{kronecker}}
}

