% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTensor_Misc.R
\name{load_orl}
\alias{load_orl}
\title{ORL Database of Faces}
\format{A Tensor object with modes 92 x 112 x 40 x 10. The first two modes correspond to the image pixels, the third mode corresponds to the individual, and the last mode correpsonds to the lighting.}
\source{
\url{https://www.kaggle.com/kasikrit/att-database-of-faces}
}
\usage{
load_orl()
}
\description{
A dataset containing pictures of 40 individuals under 10 different lightings. Each image has 92 x 112 pixels. Structured as a 4-tensor with modes 92 x 112 x 40 x 10. The data is now stored in figshare \url{https://ndownloader.figshare.com/files/28005669}
}
\references{
AT&T Laboratories Cambridge. \url{https://www.kaggle.com/kasikrit/att-database-of-faces}

F. Samaria, A. Harter, "Parameterisation of a Stochastic Model for Human Face Identification". IEEE Workshop on Applications of Computer Vision 1994.
}
\seealso{
\code{\link{plot_orl}}
}
\keyword{datasets}
