% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTensor_Class.R
\docType{methods}
\name{outerProd-methods}
\alias{outerProd}
\alias{outerProd,Tensor,Tensor-method}
\alias{outerProd-methods}
\title{Tensors Outer Product}
\usage{
outerProd(tnsr1, tnsr2)

\S4method{outerProd}{Tensor,Tensor}(tnsr1, tnsr2)
}
\arguments{
\item{tnsr1}{first Tensor instance}

\item{tnsr2}{second Tensor instance}
}
\value{
outer product between \code{tnsr1} and \code{tnsr2}
}
\description{
Returns the outer product between two Tensors
}
\details{
\code{outerProd(tnsr1,tnsr2)}
}
\examples{
tnsr1 <- rand_tensor()
tnsr2 <- rand_tensor()
outerProd(tnsr1,tnsr2)
}

