% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cs_fold}
\alias{cs_fold}
\title{Column Space Folding of Matrix}
\usage{
cs_fold(mat, m = NULL, modes = NULL)
}
\arguments{
\item{mat}{matrix to be folded}

\item{m}{the mode corresponding to cs_unfold}

\item{modes}{the original modes of the tensor}
}
\description{
DEPRECATED. Please see \code{\link{unmatvec}}
}

