% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTensor_Class.R
\docType{methods}
\name{cs_unfold-methods}
\alias{cs_unfold}
\alias{cs_unfold,Tensor-method}
\alias{cs_unfold-methods}
\title{Tensor Column Space Unfolding}
\usage{
cs_unfold(tnsr, m)

\S4method{cs_unfold}{Tensor}(tnsr, m = NULL)
}
\arguments{
\item{tnsr}{Tensor instance}

\item{m}{mode to be unfolded on}
}
\description{
DEPRECATED. Please see \code{\link{matvec-methods}} and \code{\link{unfold-methods}}.
}
\details{
\code{cs_unfold(tnsr,m=NULL)}
}

