% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{dim-methods}
\alias{dim,Tensor-method}
\alias{dim-methods}
\title{Mode Getter for Tensor}
\usage{
\S4method{dim}{Tensor}(x)
}
\arguments{
\item{x}{the Tensor instance}
}
\value{
an integer vector of the modes associated with \code{x}
}
\description{
Return the vector of modes from a tensor
}
\details{
\code{dim(x)}
}
\examples{
tnsr <- rand_tensor()
dim(tnsr)
}

