% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTensor_Misc.R
\name{plot_orl}
\alias{plot_orl}
\title{Function to plot the ORL Database of Faces}
\usage{
plot_orl(subject = 1, condition = 1)
}
\arguments{
\item{subject}{which subject to plot (1-40)}

\item{condition}{which lighting condition (1-10)}
}
\description{
A wrapper function to image() to allow easy visualization of faces_tnsr, the ORL Face Dataset.
}
\examples{
plot_orl(subject=5,condition=4)
plot_orl(subject=2,condition=7)
}
\references{
AT&T Laboratories Cambridge. \url{http://www.cl.cam.ac.uk/research/dtg/attarchive/facedatabase.html}

F. Samaria, A. Harter, "Parameterisation of a Stochastic Model for Human Face Identification". IEEE Workshop on Applications of Computer Vision 1994.
}
\seealso{
\code{\link{faces_tnsr}}
}

