% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTensor_Class.R
\docType{methods}
\name{print-methods}
\alias{print,Tensor-method}
\alias{print-methods}
\title{Print for Tensor}
\usage{
\S4method{print}{Tensor}(x, ...)
}
\arguments{
\item{x}{the Tensor instance}

\item{...}{additional parameters to be passed into print()}
}
\description{
Extend print for Tensor
}
\details{
\code{print(x,...)}
}
\examples{
tnsr <- rand_tensor()
print(tnsr)
}
\seealso{
\code{\link{show}}
}

