% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rs_fold}
\alias{rs_fold}
\title{Row Space Folding of Matrix}
\usage{
rs_fold(mat, m = NULL, modes = NULL)
}
\arguments{
\item{mat}{matrix to be folded}

\item{m}{the mode corresponding to rs_unfold}

\item{modes}{the original modes of the tensor}
}
\description{
DEPRECATED. Please see \code{\link{k_fold}}.
}

