% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTensor_Class.R
\docType{methods}
\name{rs_unfold-methods}
\alias{rs_unfold}
\alias{rs_unfold,Tensor-method}
\alias{rs_unfold-methods}
\title{Tensor Row Space Unfolding}
\usage{
rs_unfold(tnsr, m)

\S4method{rs_unfold}{Tensor}(tnsr, m = NULL)
}
\arguments{
\item{tnsr}{Tensor instance}

\item{m}{mode to be unfolded on}
}
\description{
DEPRECATED. Please see \code{\link{k_unfold-methods}} and \code{\link{unfold-methods}}.
}
\details{
\code{rs_unfold(tnsr,m=NULL)}
}

