% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTensor_Class.R
\docType{methods}
\name{show-methods}
\alias{show,Tensor-method}
\alias{show-methods}
\title{Show for Tensor}
\usage{
\S4method{show}{Tensor}(object)
}
\arguments{
\item{object}{the Tensor instance}

\item{...}{additional parameters to be passed into show()}
}
\description{
Extend show for Tensor
}
\details{
\code{show(object)}
}
\examples{
tnsr <- rand_tensor()
tnsr
}
\seealso{
\code{\link{print}}
}

