% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{vec-methods}
\alias{vec}
\alias{vec,Tensor-method}
\alias{vec-methods}
\title{Tensor Vec}
\usage{
vec(tnsr)

\S4method{vec}{Tensor}(tnsr)
}
\arguments{
\item{tnsr}{the Tensor instance}
}
\value{
vector with length \code{prod(x@modes)}
}
\description{
Turns the tensor into a single vector, following the convention that earlier indices vary slower than later indices.
}
\details{
\code{vec(tnsr)}
}
\examples{
tnsr <- rand_tensor(c(4,5,6,7))
vec(tnsr)
}
\references{
T. Kolda, B. Bader, "Tensor decomposition and applications". SIAM Applied Mathematics and Applications 2009.
}

